requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
lxml>=4.9.0
typer>=0.9.0
pyyaml>=6.0
rich>=13.0.0
pytest>=7.0.0
tkinterdnd2>=0.3.0  # Optional: enables drag & drop in GUI
customtkinter>=5.2.0  # Modern UI widgets

# AI features (optional, for auto-detect)
transformers>=4.30.0  # CLIP model for scene classification
torch>=2.0.0  # PyTorch backend for transformers
